src/agentic/refactor/requirements.txt
academy-py==0.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiologic==0.16.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
bcrypt==5.0.0
blake3==1.0.8
blinker==1.9.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
comm==0.2.3
compressed-tensors==0.11.0
cryptography==46.0.3
culsans==0.10.0
cupy-cuda12x==13.6.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
dragonhpc==0.13
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
-e git+https://github.com/stride-research/flowgentic@88a7c06c285e8b2cd47beeed1aa519cc605faaf1#egg=flowgentic
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
gguf==0.17.1
globus-sdk==3.65.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
interegular==0.3.3
invoke==2.2.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.1
langchain-mcp-adapters==0.1.14
langchain-ollama==1.0.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.23.1
mdurl==0.1.2
mistral_common==1.8.6
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.9.0
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.0
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.4.3
pycapnp==2.1.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pynvml==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
radical.asyncflow @ file:///anvil/scratch/x-mason/IMPRESS/src/agentic/refactor/radical.asyncflow
ray==2.52.1
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triton==3.4.0
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
-e git+https://github.com/radical-cybertools/aici-dragon-inference.git@6d4942de8885104c971de2add80e5517e77033fa#egg=vllm_dragonhpc
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
